+++ packages/eslint-plugin/src/rules/index.ts
@@ -1,11 +1,13 @@
 import adjacentOverloadSignatures from './adjacent-overload-signatures';
 import arrayType from './array-type';
+import awaitThenable from './await-thenable';
 import banTsIgnore from './ban-ts-ignore';
 import banTypes from './ban-types';
 import camelcase from './camelcase';
 import classNameCasing from './class-name-casing';
 import explicitFunctionReturnType from './explicit-function-return-type';
 import explicitMemberAccessibility from './explicit-member-accessibility';
+import funcCallSpacing from './func-call-spacing';
 import genericTypeNaming from './generic-type-naming';
 import indent from './indent';
 import interfaceNamePrefix from './interface-name-prefix';
@@ -16,6 +18,7 @@ import noAngleBracketTypeAssertion from './no-angle-bracket-type-assertion';
 import noArrayConstructor from './no-array-constructor';
 import noEmptyInterface from './no-empty-interface';
 import noExplicitAny from './no-explicit-any';
+import noExtraParens from './no-extra-parens';
 import noExtraneousClass from './no-extraneous-class';
 import noForInArray from './no-for-in-array';
 import noInferrableTypes from './no-inferrable-types';
@@ -34,24 +37,31 @@ import noUnusedVars from './no-unused-vars';
 import noUseBeforeDefine from './no-use-before-define';
 import noUselessConstructor from './no-useless-constructor';
 import noVarRequires from './no-var-requires';
+import preferForOf from './prefer-for-of';
 import preferFunctionType from './prefer-function-type';
+import preferIncludes from './prefer-includes';
 import preferInterface from './prefer-interface';
 import preferNamespaceKeyword from './prefer-namespace-keyword';
+import preferStringStartsEndsWith from './prefer-string-starts-ends-with';
 import promiseFunctionAsync from './promise-function-async';
 import requireArraySortCompare from './require-array-sort-compare';
 import restrictPlusOperands from './restrict-plus-operands';
+import semi from './semi';
 import typeAnnotationSpacing from './type-annotation-spacing';
+import unboundMethod from './unbound-method';
 import unifiedSignatures from './unified-signatures';
 
 export default {
   'adjacent-overload-signatures': adjacentOverloadSignatures,
   'array-type': arrayType,
+  'await-thenable': awaitThenable,
   'ban-ts-ignore': banTsIgnore,
   'ban-types': banTypes,
   camelcase: camelcase,
   'class-name-casing': classNameCasing,
   'explicit-function-return-type': explicitFunctionReturnType,
   'explicit-member-accessibility': explicitMemberAccessibility,
+  'func-call-spacing': funcCallSpacing,
   'generic-type-naming': genericTypeNaming,
   indent: indent,
   'interface-name-prefix': interfaceNamePrefix,
@@ -62,6 +72,7 @@ export default {
   'no-array-constructor': noArrayConstructor,
   'no-empty-interface': noEmptyInterface,
   'no-explicit-any': noExplicitAny,
+  'no-extra-parens': noExtraParens,
   'no-extraneous-class': noExtraneousClass,
   'no-for-in-array': noForInArray,
   'no-inferrable-types': noInferrableTypes,
@@ -80,12 +91,17 @@ export default {
   'no-use-before-define': noUseBeforeDefine,
   'no-useless-constructor': noUselessConstructor,
   'no-var-requires': noVarRequires,
+  'prefer-for-of': preferForOf,
   'prefer-function-type': preferFunctionType,
+  'prefer-includes': preferIncludes,
   'prefer-interface': preferInterface,
   'prefer-namespace-keyword': preferNamespaceKeyword,
+  'prefer-string-starts-ends-with': preferStringStartsEndsWith,
   'promise-function-async': promiseFunctionAsync,
   'require-array-sort-compare': requireArraySortCompare,
   'restrict-plus-operands': restrictPlusOperands,
+  semi: semi,
   'type-annotation-spacing': typeAnnotationSpacing,
+  'unbound-method': unboundMethod,
   'unified-signatures': unifiedSignatures,
 };