+++ packages/eslint-plugin/src/util/index.ts
@@ -1,4 +1,5 @@
 export * from './applyDefault';
+export * from './astUtils';
 export * from './createRule';
 export * from './deepMerge';
 export * from './getParserServices';