+++ packages/eslint-plugin/tests/rules/indent.test.ts
@@ -742,6 +742,30 @@ const foo : Foo = {
             `,
       options: [4, { VariableDeclarator: { const: 3 } }],
     },
+    {
+      code: `
+const name: string = '  Typescript  '
+        .toUpperCase()
+        .trim(),
+
+      greeting: string = (" Hello " + name)
+        .toUpperCase()
+        .trim();
+            `,
+      options: [2, { VariableDeclarator: { const: 3 } }],
+    },
+    {
+      code: `
+const div: JQuery<HTMLElement> = $('<div>')
+        .addClass('some-class')
+        .appendTo($('body')),
+
+      button: JQuery<HTMLElement> = $('<button>')
+        .text('Cancel')
+        .appendTo(div);
+            `,
+      options: [2, { VariableDeclarator: { const: 3 } }],
+    },
   ],
   invalid: [
     ...individualNodeTests.invalid,